gcp/data-governance-workflows-dataplex-workflows/code/terraform/function_code/main.py
"""
Data Quality Assessment Cloud Function for Governance Workflows
This function performs automated data quality assessments using Dataplex and BigQuery.
"""

import functions_framework
from google.cloud import bigquery
from google.cloud import dataplex_v1
import json
import datetime
import logging
import os
from typing import Dict, Any, Optional

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

# Environment variables
PROJECT_ID = os.environ.get('PROJECT_ID', '${project_id}')
DATASET_NAME = os.environ.get('DATASET_NAME', '${dataset_name}')

@functions_framework.http
def assess_data_quality(request):
    """
    Assess data quality for governance workflows
    
    Args:
        request: Flask request object containing assessment parameters
        
    Returns:
        JSON response with quality assessment results
    """
    try:
        # Initialize clients
        bigquery_client = bigquery.Client(project=PROJECT_ID)
        dataplex_client = dataplex_v1.DataplexServiceClient()
        
        # Parse request parameters
        request_json = request.get_json(silent=True)
        if not request_json:
            return {
                'status': 'error',
                'message': 'No JSON data provided in request'
            }, 400
        
        asset_name = request_json.get('asset_name', 'unknown')
        zone = request_json.get('zone', 'unknown')
        
        logger.info(f"Assessing data quality for asset: {asset_name}, zone: {zone}")
        
        # Perform comprehensive quality assessment
        quality_metrics = assess_data_quality_metrics(asset_name, zone)
        
        # Calculate overall quality score
        quality_score = calculate_overall_quality_score(quality_metrics)
        
        # Count issues found
        issues_found = count_quality_issues(quality_metrics)
        
        # Store results in BigQuery
        storage_result = store_quality_results(
            bigquery_client,
            asset_name,
            zone,
            quality_score,
            issues_found,
            quality_metrics
        )
        
        # Prepare response
        response = {
            'status': 'success',
            'asset_name': asset_name,
            'zone': zone,
            'quality_score': quality_score,
            'issues_found': issues_found,
            'metrics': quality_metrics,
            'timestamp': datetime.datetime.utcnow().isoformat(),
            'storage_result': storage_result
        }
        
        logger.info(f"Quality assessment completed for {asset_name}: score={quality_score}")
        return response
        
    except Exception as e:
        logger.error(f"Error in data quality assessment: {str(e)}")
        return {
            'status': 'error',
            'message': str(e),
            'timestamp': datetime.datetime.utcnow().isoformat()
        }, 500

def assess_data_quality_metrics(asset_name: str, zone: str) -> Dict[str, float]:
    """
    Assess various data quality metrics for the given asset
    
    Args:
        asset_name: Name of the data asset
        zone: Zone where the asset is located
        
    Returns:
        Dictionary containing quality metrics
    """
    try:
        # In a real implementation, this would connect to actual data sources
        # and perform comprehensive quality checks
        
        # Simulate quality metrics based on asset characteristics
        base_scores = {
            'completeness': 0.95,    # Percentage of non-null values
            'validity': 0.88,        # Percentage of valid values
            'consistency': 0.92,     # Consistency across related fields
            'accuracy': 0.89,        # Accuracy against reference data
            'timeliness': 0.94,      # Data freshness and update frequency
            'uniqueness': 0.91       # Percentage of unique values where expected
        }
        
        # Adjust scores based on zone type
        if zone == 'raw-data-zone':
            # Raw data typically has lower quality scores
            adjustment_factor = 0.95
        elif zone == 'curated-data-zone':
            # Curated data should have higher quality scores
            adjustment_factor = 1.05
        else:
            adjustment_factor = 1.0
        
        # Apply zone-based adjustments
        adjusted_scores = {
            metric: min(1.0, score * adjustment_factor)
            for metric, score in base_scores.items()
        }
        
        # Add asset-specific variations
        if 'customer' in asset_name.lower():
            adjusted_scores['accuracy'] *= 0.98  # Customer data needs high accuracy
        elif 'transaction' in asset_name.lower():
            adjusted_scores['timeliness'] *= 1.02  # Transaction data needs timeliness
        
        return adjusted_scores
        
    except Exception as e:
        logger.error(f"Error assessing quality metrics: {str(e)}")
        # Return default scores on error
        return {
            'completeness': 0.8,
            'validity': 0.8,
            'consistency': 0.8,
            'accuracy': 0.8,
            'timeliness': 0.8,
            'uniqueness': 0.8
        }

def calculate_overall_quality_score(metrics: Dict[str, float]) -> float:
    """
    Calculate overall quality score from individual metrics
    
    Args:
        metrics: Dictionary of quality metrics
        
    Returns:
        Overall quality score between 0.0 and 1.0
    """
    try:
        # Define weights for different metrics
        weights = {
            'completeness': 0.2,
            'validity': 0.2,
            'consistency': 0.15,
            'accuracy': 0.2,
            'timeliness': 0.15,
            'uniqueness': 0.1
        }
        
        # Calculate weighted average
        total_score = 0.0
        total_weight = 0.0
        
        for metric, score in metrics.items():
            if metric in weights:
                total_score += score * weights[metric]
                total_weight += weights[metric]
        
        # Normalize score
        if total_weight > 0:
            return round(total_score / total_weight, 3)
        else:
            return 0.0
            
    except Exception as e:
        logger.error(f"Error calculating overall quality score: {str(e)}")
        return 0.0

def count_quality_issues(metrics: Dict[str, float]) -> int:
    """
    Count the number of quality issues based on metrics
    
    Args:
        metrics: Dictionary of quality metrics
        
    Returns:
        Number of quality issues found
    """
    try:
        # Define thresholds for quality issues
        thresholds = {
            'completeness': 0.95,
            'validity': 0.90,
            'consistency': 0.85,
            'accuracy': 0.90,
            'timeliness': 0.85,
            'uniqueness': 0.90
        }
        
        issues_count = 0
        for metric, score in metrics.items():
            if metric in thresholds and score < thresholds[metric]:
                issues_count += 1
        
        return issues_count
        
    except Exception as e:
        logger.error(f"Error counting quality issues: {str(e)}")
        return 0

def store_quality_results(
    client: bigquery.Client,
    asset_name: str,
    zone: str,
    quality_score: float,
    issues_found: int,
    metrics: Dict[str, float]
) -> Dict[str, Any]:
    """
    Store quality assessment results in BigQuery
    
    Args:
        client: BigQuery client
        asset_name: Name of the assessed asset
        zone: Zone where the asset is located
        quality_score: Overall quality score
        issues_found: Number of issues found
        metrics: Detailed quality metrics
        
    Returns:
        Dictionary with storage result information
    """
    try:
        # Prepare the row data
        table_id = f"{PROJECT_ID}.{DATASET_NAME}.data_quality_metrics"
        
        row_data = {
            'timestamp': datetime.datetime.utcnow().isoformat(),
            'asset_name': asset_name,
            'zone': zone,
            'quality_score': quality_score,
            'issues_found': issues_found,
            'scan_type': 'automated',
            'details': json.dumps(metrics)
        }
        
        # Insert the row
        errors = client.insert_rows_json(table_id, [row_data])
        
        if errors:
            logger.error(f"Error inserting row into BigQuery: {errors}")
            return {
                'success': False,
                'errors': errors
            }
        else:
            logger.info(f"Successfully stored quality results for {asset_name}")
            return {
                'success': True,
                'table_id': table_id,
                'row_count': 1
            }
            
    except Exception as e:
        logger.error(f"Error storing quality results: {str(e)}")
        return {
            'success': False,
            'error': str(e)
        }

def get_asset_metadata(asset_name: str, zone: str) -> Dict[str, Any]:
    """
    Get metadata for a data asset from Dataplex
    
    Args:
        asset_name: Name of the data asset
        zone: Zone where the asset is located
        
    Returns:
        Dictionary containing asset metadata
    """
    try:
        # In a real implementation, this would query Dataplex APIs
        # for actual asset metadata
        
        # Simulate metadata based on asset characteristics
        metadata = {
            'asset_type': 'TABLE' if 'table' in asset_name.lower() else 'BUCKET',
            'size_bytes': 1024 * 1024 * 10,  # 10MB example
            'row_count': 1000,
            'column_count': 15,
            'last_modified': datetime.datetime.utcnow().isoformat(),
            'data_format': 'JSON' if 'json' in asset_name.lower() else 'CSV',
            'compression': 'GZIP',
            'encryption': 'CUSTOMER_MANAGED'
        }
        
        return metadata
        
    except Exception as e:
        logger.error(f"Error getting asset metadata: {str(e)}")
        return {}

def validate_data_schema(asset_name: str, zone: str) -> Dict[str, Any]:
    """
    Validate the data schema for consistency and standards compliance
    
    Args:
        asset_name: Name of the data asset
        zone: Zone where the asset is located
        
    Returns:
        Dictionary containing schema validation results
    """
    try:
        # In a real implementation, this would perform actual schema validation
        
        # Simulate schema validation results
        validation_results = {
            'schema_valid': True,
            'column_names_standard': True,
            'data_types_consistent': True,
            'missing_columns': [],
            'extra_columns': [],
            'type_mismatches': []
        }
        
        return validation_results
        
    except Exception as e:
        logger.error(f"Error validating data schema: {str(e)}")
        return {
            'schema_valid': False,
            'error': str(e)
        }

# Health check endpoint
@functions_framework.http
def health_check(request):
    """
    Health check endpoint for the function
    
    Returns:
        JSON response indicating function health
    """
    return {
        'status': 'healthy',
        'timestamp': datetime.datetime.utcnow().isoformat(),
        'function': 'data-quality-assessor',
        'version': '1.0.0'
    }