gcp/data-governance-workflows-dataplex-workflows/code/terraform/function_code/requirements.txt
# Requirements for Data Quality Assessment Cloud Function

# Google Cloud Functions Framework
functions-framework==3.*

# Google Cloud client libraries
google-cloud-bigquery==3.*
google-cloud-dataplex==1.*
google-cloud-storage==2.*
google-cloud-dlp==3.*

# Additional dependencies for data processing
pandas==2.*
numpy==1.*

# HTTP and JSON handling
requests==2.*

# Logging and monitoring
google-cloud-logging==3.*
google-cloud-monitoring==2.*

# Date and time handling
python-dateutil==2.*

# JSON schema validation
jsonschema==4.*

# Cryptography for security
cryptography==41.*